+++ tests/Tests/Data/Annotations/SingleTestClass.php
@@ -20,6 +20,9 @@
 
 namespace AppserverIo\Doppelgaenger\Tests\Data\Annotations;
 
+use AppserverIo\Doppelgaenger\Exceptions\BrokenInvariantException;
+use AppserverIo\Doppelgaenger\Tests\Data\Aspects\JustAdvicesAspect;
+
 /**
  * Test class containing directly annotated methods with one advice each, covering all possible join-points
  *
@@ -35,7 +38,7 @@ class SingleTestClass
     /**
      * @Before("weave(Logger->error($param1))")
      */
-    public function iHaveABeforeAdvice($param1)
+    public function iHaveABeforeWeave($param1)
     {
 
     }
@@ -43,7 +46,7 @@ class SingleTestClass
     /**
      * @After("weave(Logger->error($param1))")
      */
-    public function iHaveAnAfterAdvice($param1)
+    public function iHaveAnAfterWeave($param1)
     {
 
     }
@@ -51,13 +54,47 @@ class SingleTestClass
     /**
      * @AfterThrowing("weave(Logger->error($param1))")
      */
-    public function iHaveAnAfterThrowingAdvice($param1)
+    public function iHaveAnAfterThrowingWeave($param1)
     {
 
     }
 
     /**
      * @AfterReturning("weave(Logger->error($param1))")
+     */
+    public function iHaveAnAfterReturningWeave($param1)
+    {
+
+    }
+
+    /**
+     * @Before("advise(JustAdvicesAspect->basicBeforeAdvice())")
+     */
+    public function iHaveABeforeAdvice($param1)
+    {
+
+    }
+
+    /**
+     * @After("advise(JustAdvicesAspect->basicAfterAdvice())")
+     */
+    public function iHaveAnAfterAdvice($param1)
+    {
+
+    }
+
+    /**
+     * We throw a very weird exception so we can test if we get the right instance in our advice
+     *
+     * @AfterThrowing("advise(JustAdvicesAspect->basicAfterThrowingAdvice())")
+     */
+    public function iHaveAnAfterThrowingAdvice($param1)
+    {
+        throw new BrokenInvariantException(__FUNCTION__);
+    }
+
+    /**
+     * @AfterReturning("advise(JustAdvicesAspect->basicAfterReturningAdvice())")
      */
     public function iHaveAnAfterReturningAdvice($param1)
     {